requirements/base.txt
Django==1.7
django-braces==1.4.0
Pillow==2.6.0
sqlparse==0.1.12
